+++ playground/dynamic-import/views/Two.vue
@@ -1,0 +1,3 @@
+<template>
+  <h3 class="dynamic-import-two">Two</h3>
+</template>