+++ src/node/utils/shims.d.ts
@@ -1,0 +1,3 @@
+declare module 'merge-source-map' {
+  export default function merge(oldMap: object, newMap: object): object
+}